chore: add themes files to exports map, and rename bundle to index.js
